Adicionado o padrão visitor para a resolução da problemática de imprimir uma árvore de expressões.

+++ DesignPatterns/Interpreter/Calculadora/Program.cs
@@ -6,6 +6,6 @@ using Calculadora.Models;
 IExpressao esquerda = new Multiplicacao(new Numero(4), new Soma(new Numero(22), new Numero(3)));
 IExpressao direita = new Multiplicacao(new Numero(10), new Numero(5));
 
-int resultado = new Subtracao(esquerda, direita).Avalia();
+double resultado = new Subtracao(esquerda, direita).Avalia();
 
 Console.WriteLine($"O valor final da expressão é {resultado}");
